Moved core into core module.
